+++ lib/src/schema/all.dart
@@ -91,3 +91,4 @@ export 'schemas/text_form_field_schema.dart';
 export 'schemas/text_schema.dart';
 export 'schemas/theme_schema.dart';
 export 'schemas/tween_animation_schema.dart';
+export 'schemas/wrap_schema.dart';