+++ lib/src/schema/schemas/json_widget_data_schema.dart
@@ -27,8 +27,15 @@ class JsonWidgetDataSchema {
           },
           'child': SchemaHelper.anySchema,
           'children': {
-            'type': 'array',
-            'items': SchemaHelper.anySchema,
+            'anyOf': [
+              {
+                'type': 'string',
+              },
+              {
+                'type': 'array',
+                'items': SchemaHelper.anySchema,
+              },
+            ],
           },
           'id': {
             'type': 'string',